+++ skeleton_extract/Point3D.h
@@ -2,10 +2,10 @@
 class Point3D
 {
 public:
-	double x,y,z;
+	float x,y,z;
 	Point3D(void);
 	~Point3D(void);
-	Point3D(double x0,double y0,double z0);
+	Point3D(float x0,float y0,float z0);
 	Point3D(const Point3D& p);
 	//double dist(const Point3D& p);
 };
